EvoNN-Prism/pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Family-based evolutionary neural architecture search for Apple Silicon"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.1.3",
    "mlx>=0.31.0",
    "numpy>=2.2.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "typer>=0.15.1",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]
viz = ["matplotlib>=3.9"]
data = ["openml>=0.14", "pandas>=2.2"]
baselines = ["xgboost>=2.0", "lightgbm>=4.0"]
all = ["prism[dev,viz,data,baselines]"]

[project.scripts]
prism = "prism.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prism"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["prism"]

[tool.coverage.report]
fail_under = 40
show_missing = true
skip_covered = true
